moved Network(View)Client(Messages/Responses) to near_client (#7908)
These actix messages are an implementation detail of near_client crate.
