+++ chain/client/src/tests/catching_up.rs
@@ -6,6 +6,7 @@ use actix::{Addr, System};
 use borsh::{BorshDeserialize, BorshSerialize};
 use futures::{future, FutureExt};
 
+use crate::adapter::NetworkClientMessages;
 use crate::test_utils::setup_mock_all_validators;
 use crate::{ClientActor, Query, ViewClientActor};
 use near_actix_test_utils::run_actix;
@@ -13,9 +14,7 @@ use near_chain::test_utils::{account_id_to_shard_id, ValidatorSchedule};
 use near_chain_configs::TEST_STATE_SYNC_TIMEOUT;
 use near_crypto::{InMemorySigner, KeyType};
 use near_network::types::{AccountIdOrPeerTrackingShard, AccountOrPeerIdOrHash, PeerInfo};
-use near_network::types::{
-    NetworkClientMessages, NetworkRequests, NetworkResponses, PeerManagerMessageRequest,
-};
+use near_network::types::{NetworkRequests, NetworkResponses, PeerManagerMessageRequest};
 use near_o11y::testonly::init_integration_logger;
 use near_o11y::WithSpanContextExt;
 use near_primitives::hash::{hash as hash_func, CryptoHash};
