+++ chain/client/src/tests/consensus.rs
@@ -5,14 +5,13 @@ use actix::{Addr, System};
 use near_chain::test_utils::ValidatorSchedule;
 use rand::{thread_rng, Rng};
 
+use crate::adapter::NetworkClientMessages;
 use crate::test_utils::setup_mock_all_validators;
 use crate::{ClientActor, ViewClientActor};
 use near_actix_test_utils::run_actix;
 use near_chain::Block;
 use near_network::types::PeerInfo;
-use near_network::types::{
-    NetworkClientMessages, NetworkRequests, NetworkResponses, PeerManagerMessageRequest,
-};
+use near_network::types::{NetworkRequests, NetworkResponses, PeerManagerMessageRequest};
 use near_o11y::testonly::init_integration_logger;
 use near_o11y::WithSpanContextExt;
 use near_primitives::block::{Approval, ApprovalInner};
