+++ chain/client/src/tests/cross_shard_tx.rs
@@ -7,13 +7,13 @@ use std::sync::{Arc, RwLock};
 use actix::{Addr, MailboxError, System};
 use futures::{future, FutureExt};
 
+use crate::adapter::{NetworkClientMessages, NetworkClientResponses};
 use near_actix_test_utils::run_actix;
 use near_chain::test_utils::{account_id_to_shard_id, ValidatorSchedule};
 use near_crypto::{InMemorySigner, KeyType};
 use near_network::types::PeerInfo;
 use near_network::types::{
-    NetworkClientMessages, NetworkClientResponses, NetworkResponses, PeerManagerMessageRequest,
-    PeerManagerMessageResponse,
+    NetworkResponses, PeerManagerMessageRequest, PeerManagerMessageResponse,
 };
 use near_o11y::testonly::init_integration_logger;
 use near_o11y::WithSpanContextExt;
