+++ chain/network/src/types.rs
@@ -9,16 +9,14 @@ use futures::future::BoxFuture;
 use futures::FutureExt;
 use near_crypto::PublicKey;
 use near_o11y::WithSpanContext;
-use near_primitives::block::{Approval, ApprovalMessage, Block, BlockHeader};
+use near_primitives::block::{ApprovalMessage, Block};
 use near_primitives::challenge::Challenge;
-use near_primitives::errors::InvalidTxError;
 use near_primitives::hash::CryptoHash;
 use near_primitives::network::{AnnounceAccount, PeerId};
-use near_primitives::sharding::{PartialEncodedChunk, PartialEncodedChunkWithArcReceipts};
+use near_primitives::sharding::PartialEncodedChunkWithArcReceipts;
 use near_primitives::transaction::SignedTransaction;
 use near_primitives::types::BlockHeight;
 use near_primitives::types::{AccountId, EpochId, ShardId};
-use near_primitives::views::FinalExecutionOutcomeView;
 use near_primitives::views::{KnownProducerView, NetworkInfoView, PeerInfoView};
 use once_cell::sync::OnceCell;
 use std::collections::HashMap;
@@ -397,67 +395,17 @@ pub enum NetworkResponses {
     RouteNotFound,
 }
 
-#[derive(actix::Message, Debug, strum::AsRefStr, strum::IntoStaticStr)]
-// TODO(#1313): Use Box
-#[allow(clippy::large_enum_variant)]
-#[rtype(result = "NetworkClientResponses")]
-pub enum NetworkClientMessages {
-    #[cfg(feature = "test_features")]
-    Adversarial(crate::types::NetworkAdversarialMessage),
-
-    /// Received transaction.
-    Transaction {
-        transaction: SignedTransaction,
-        /// Whether the transaction is forwarded from other nodes.
-        is_forwarded: bool,
-        /// Whether the transaction needs to be submitted.
-        check_only: bool,
-    },
-    /// Received block, possibly requested.
-    Block(Block, PeerId, bool),
-    /// Received list of headers for syncing.
-    BlockHeaders(Vec<BlockHeader>, PeerId),
-    /// Block approval.
-    BlockApproval(Approval, PeerId),
-    /// State response.
-    StateResponse(StateResponseInfo),
-
-    /// Request chunk parts and/or receipts.
-    PartialEncodedChunkRequest(PartialEncodedChunkRequestMsg, CryptoHash),
-    /// Response to a request for  chunk parts and/or receipts.
-    PartialEncodedChunkResponse(PartialEncodedChunkResponseMsg, std::time::Instant),
-    /// Information about chunk such as its header, some subset of parts and/or incoming receipts
-    PartialEncodedChunk(PartialEncodedChunk),
-    /// Forwarding parts to those tracking the shard (so they don't need to send requests)
-    PartialEncodedChunkForward(PartialEncodedChunkForwardMsg),
-
-    /// A challenge to invalidate the block.
-    Challenge(Challenge),
-
-    NetworkInfo(NetworkInfo),
-}
-
-// TODO(#1313): Use Box
-#[derive(Eq, PartialEq, Debug, actix::MessageResponse)]
-#[allow(clippy::large_enum_variant)]
-pub enum NetworkClientResponses {
-    /// Adv controls.
-    #[cfg(feature = "test_features")]
-    AdvResult(u64),
-
-    /// No response.
-    NoResponse,
-    /// Valid transaction inserted into mempool as response to Transaction.
-    ValidTx,
-    /// Invalid transaction inserted into mempool as response to Transaction.
-    InvalidTx(InvalidTxError),
-    /// The request is routed to other shards
-    RequestRouted,
-    /// The node being queried does not track the shard needed and therefore cannot provide userful
-    /// response.
-    DoesNotTrackShard,
-    /// Ban peer for malicious behavior.
-    Ban { ban_reason: ReasonForBan },
+#[cfg(feature = "test_features")]
+#[derive(actix::Message, Debug)]
+#[rtype(result = "Option<u64>")]
+pub enum NetworkAdversarialMessage {
+    AdvProduceBlocks(u64, bool),
+    AdvSwitchToHeight(u64),
+    AdvDisableHeaderSync,
+    AdvDisableDoomslug,
+    AdvGetSavedBlocks,
+    AdvCheckStorageConsistency,
+    AdvSetSyncInfo(u64),
 }
 
 pub trait MsgRecipient<M: actix::Message>: Send + Sync + 'static {
@@ -479,7 +427,6 @@ where
         actix::Addr::do_send(self, msg)
     }
 }
-
 pub trait PeerManagerAdapter:
     MsgRecipient<WithSpanContext<PeerManagerMessageRequest>>
     + MsgRecipient<WithSpanContext<SetChainInfo>>
@@ -542,8 +489,6 @@ mod tests {
         assert_size!(HandshakeFailureReason);
         assert_size!(NetworkRequests);
         assert_size!(NetworkResponses);
-        assert_size!(NetworkClientMessages);
-        assert_size!(NetworkClientResponses);
         assert_size!(Handshake);
         assert_size!(Ping);
         assert_size!(Pong);
@@ -635,58 +580,4 @@ pub struct AccountIdOrPeerTrackingShard {
     pub only_archival: bool,
     /// Only send messages to peers whose latest chain height is no less `min_height`
     pub min_height: BlockHeight,
-}
-
-#[derive(actix::Message, strum::IntoStaticStr)]
-#[rtype(result = "NetworkViewClientResponses")]
-pub enum NetworkViewClientMessages {
-    #[cfg(feature = "test_features")]
-    Adversarial(NetworkAdversarialMessage),
-
-    /// Transaction status query
-    TxStatus { tx_hash: CryptoHash, signer_account_id: AccountId },
-    /// Transaction status response
-    TxStatusResponse(Box<FinalExecutionOutcomeView>),
-    /// Request a block.
-    BlockRequest(CryptoHash),
-    /// Request headers.
-    BlockHeadersRequest(Vec<CryptoHash>),
-    /// State request header.
-    StateRequestHeader { shard_id: ShardId, sync_hash: CryptoHash },
-    /// State request part.
-    StateRequestPart { shard_id: ShardId, sync_hash: CryptoHash, part_id: u64 },
-    /// Account announcements that needs to be validated before being processed.
-    /// They are paired with last epoch id known to this announcement, in order to accept only
-    /// newer announcements.
-    AnnounceAccount(Vec<(AnnounceAccount, Option<EpochId>)>),
-}
-
-#[derive(Debug, actix::MessageResponse)]
-pub enum NetworkViewClientResponses {
-    /// Transaction execution outcome
-    TxStatus(Box<FinalExecutionOutcomeView>),
-    /// Block response.
-    Block(Box<Block>),
-    /// Headers response.
-    BlockHeaders(Vec<BlockHeader>),
-    /// Response to state request.
-    StateResponse(Box<StateResponseInfo>),
-    /// Valid announce accounts.
-    AnnounceAccount(Vec<AnnounceAccount>),
-    /// Ban peer for malicious behavior.
-    Ban { ban_reason: ReasonForBan },
-    /// Response not needed
-    NoResponse,
-}
-
-#[cfg(feature = "test_features")]
-#[derive(Debug)]
-pub enum NetworkAdversarialMessage {
-    AdvProduceBlocks(u64, bool),
-    AdvSwitchToHeight(u64),
-    AdvDisableHeaderSync,
-    AdvDisableDoomslug,
-    AdvGetSavedBlocks,
-    AdvCheckStorageConsistency,
-    AdvSetSyncInfo(u64),
 }
