+++ chain/rosetta-rpc/src/lib.rs
@@ -741,7 +741,7 @@ async fn construction_submit(
     let transaction_hash = signed_transaction.as_ref().get_hash();
     let transaction_submittion = client_addr
         .send(
-            near_network::types::NetworkClientMessages::Transaction {
+            near_client::adapter::NetworkClientMessages::Transaction {
                 transaction: signed_transaction.into_inner(),
                 is_forwarded: false,
                 check_only: false,
@@ -750,15 +750,15 @@ async fn construction_submit(
         )
         .await?;
     match transaction_submittion {
-        near_network::types::NetworkClientResponses::ValidTx
-        | near_network::types::NetworkClientResponses::RequestRouted => {
+        near_client::adapter::NetworkClientResponses::ValidTx
+        | near_client::adapter::NetworkClientResponses::RequestRouted => {
             Ok(Json(models::TransactionIdentifierResponse {
                 transaction_identifier: models::TransactionIdentifier::transaction(
                     &transaction_hash,
                 ),
             }))
         }
-        near_network::types::NetworkClientResponses::InvalidTx(error) => {
+        near_client::adapter::NetworkClientResponses::InvalidTx(error) => {
             Err(errors::ErrorKind::InvalidInput(error.to_string()).into())
         }
         _ => Err(errors::ErrorKind::InternalInvariantError(format!(
