+++ integration-tests/src/tests/client/features/account_id_in_function_call_permission.rs
@@ -1,8 +1,8 @@
 use near_chain::{ChainGenesis, RuntimeAdapter};
 use near_chain_configs::Genesis;
+use near_client::adapter::NetworkClientResponses;
 use near_client::test_utils::TestEnv;
 use near_crypto::{InMemorySigner, KeyType, Signer};
-use near_network::types::NetworkClientResponses;
 use near_primitives::account::{AccessKey, AccessKeyPermission, FunctionCallPermission};
 use near_primitives::errors::{ActionsValidationError, InvalidTxError};
 use near_primitives::hash::CryptoHash;
