+++ tools/mirror/src/lib.rs
@@ -2,6 +2,7 @@ use actix::Addr;
 use anyhow::Context;
 use borsh::{BorshDeserialize, BorshSerialize};
 use near_chain_configs::GenesisValidationMode;
+use near_client::adapter::{NetworkClientMessages, NetworkClientResponses};
 use near_client::{ClientActor, ViewClientActor};
 use near_client_primitives::types::{
     GetBlock, GetBlockError, GetChunk, GetChunkError, GetExecutionOutcome,
@@ -9,7 +10,6 @@ use near_client_primitives::types::{
 };
 use near_crypto::{PublicKey, SecretKey};
 use near_indexer::{Indexer, StreamerMessage};
-use near_network::types::{NetworkClientMessages, NetworkClientResponses};
 use near_o11y::WithSpanContextExt;
 use near_primitives::hash::CryptoHash;
 use near_primitives::transaction::{
