+++ tools/mock-node/src/setup.rs
@@ -8,9 +8,10 @@ use near_chain::{
     Chain, ChainGenesis, ChainStore, ChainStoreAccess, DoomslugThresholdMode, RuntimeAdapter,
 };
 use near_chain_configs::GenesisConfig;
+use near_client::adapter::NetworkClientMessages;
 use near_client::{start_client, start_view_client, ClientActor, ViewClientActor};
 use near_epoch_manager::{EpochManager, EpochManagerAdapter};
-use near_network::types::{NetworkClientMessages, NetworkRecipient};
+use near_network::types::NetworkRecipient;
 use near_o11y::WithSpanContext;
 use near_primitives::state_part::PartId;
 use near_primitives::syncing::get_num_state_parts;
@@ -310,10 +311,10 @@ mod tests {
     use futures::{future, FutureExt};
     use near_actix_test_utils::{run_actix, spawn_interruptible};
     use near_chain_configs::Genesis;
+    use near_client::adapter::NetworkClientMessages;
     use near_client::GetBlock;
     use near_crypto::{InMemorySigner, KeyType};
     use near_network::test_utils::{open_port, WaitOrTimeoutActor};
-    use near_network::types::NetworkClientMessages;
     use near_o11y::testonly::init_integration_logger;
     use near_o11y::WithSpanContextExt;
     use near_primitives::hash::CryptoHash;
