+++ tools/state-viewer/src/apply_chunk.rs
@@ -407,10 +407,10 @@ pub(crate) fn apply_receipt(
 mod test {
     use near_chain::{ChainGenesis, ChainStore, ChainStoreAccess, Provenance};
     use near_chain_configs::Genesis;
+    use near_client::adapter::NetworkClientResponses;
     use near_client::test_utils::TestEnv;
     use near_crypto::{InMemorySigner, KeyType};
     use near_epoch_manager::EpochManagerAdapter;
-    use near_network::types::NetworkClientResponses;
     use near_primitives::hash::CryptoHash;
     use near_primitives::runtime::config_store::RuntimeConfigStore;
     use near_primitives::shard_layout;
